python-api/requirements.txt
pandas~=1.2.3
Flask~=1.1.2
flask-marshmallow~=0.14.0
Flask-SQLAlchemy~=2.5.1
mariadb~=1.0.6
marshmallow-sqlalchemy~=0.24.2
python-dotenv~=0.16.0
